rename: rename package location
