+++ back-end/src/main/java/sjtu/dolo/CouseApplication.java
@@ -1,11 +1,11 @@
-package sjtu.couse;
+package sjtu.dolo;
 
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-@MapperScan("sjtu.couse.mapper")
+@MapperScan("sjtu.dolo.mapper")
 public class CouseApplication {
 
 	public static void main(String[] args) {
