+++ back-end/src/main/java/sjtu/dolo/controller/StudentController.java
@@ -1,4 +1,4 @@
-package sjtu.couse.controller;
+package sjtu.dolo.controller;
 
 public class StudentController {
 }
