+++ back-end/src/main/java/sjtu/dolo/daoimpl/StudentDaoImpl.java
@@ -1,4 +1,4 @@
-package sjtu.couse.daoimpl;
+package sjtu.dolo.daoimpl;
 
 public class StudentDaoImpl {
 }
