+++ back-end/src/main/java/sjtu/dolo/mapper/AdminMapper.java
@@ -1,4 +1,4 @@
-package sjtu.couse.mapper;
+package sjtu.dolo.mapper;
 
 public interface AdminMapper {
