+++ back-end/src/main/java/sjtu/dolo/mapper/TeacherDao.java
@@ -1,4 +1,4 @@
-package sjtu.couse.mapper;
+package sjtu.dolo.mapper;
 
 public interface TeacherDao {
 }
