+++ back-end/src/main/java/sjtu/dolo/model/Course.java
@@ -1,4 +1,4 @@
-package sjtu.couse.model;
+package sjtu.dolo.model;
 
 public class Course {
     private String courseID;
