+++ back-end/src/main/java/sjtu/dolo/model/Takes.java
@@ -1,4 +1,4 @@
-package sjtu.couse.model;
+package sjtu.dolo.model;
 
 import java.math.BigDecimal;
