+++ back-end/src/main/java/sjtu/dolo/model/Teach.java
@@ -1,4 +1,4 @@
-package sjtu.couse.model;
+package sjtu.dolo.model;
 
 public class Teach {
     private String secID;
