+++ back-end/src/test/java/sjtu/dolo/CouseApplicationTests.java
@@ -1,4 +1,4 @@
-package sjtu.couse;
+package sjtu.dolo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
